✏️ Rename: `COACHES.ts` -> `COACHING_PJ.ts`
