+++ src/lib/components/coaching/Coaches.svelte
@@ -1,6 +1,6 @@
 <!-- © 2022 REVATI -->
 <script lang="ts">
-	import { COACHES } from '$lib/scripts/data/COACHES';
+	import { COACHES } from '$lib/scripts/data/COACHING_PJ';
 </script>
 
 <ul>
